Remove jmespath from requirements.txt
jmespath is helpful for paginating results from boto, but it's not implemented in mozdef_client.
Having it listed as a requirement produces a knock-on issue for setuptools' entry_points, which looks for requirements and needs them to be installed.  So jmespath needs to be either added to setup.py, or removed completely, and removing makes more sense.